Increase IOSIZE in hellworld and warpdrive cases
This commit increases the IOSIZE to 200000 (2e5) in the
helloworld and warpdive cases in order to show off the true
performance of the parallel filesystem and MPI-IO.

The prior value of this parameter was 2e4, but during
ATPESC2021 this proved inadequate to demonstrate the
intended performance on theta.alcf.anl.gov.  In particular,
the warpdrive cases both wrote data at 225-250 MiB/s, with
the non-collective warpdriveB the faster of the two.

Increasing IOSIZE increases I/O performance substantially,
with warpdriveA writing at 1300MiB/s and warpdriveB writing
at 400MiB/s.  A similar change to the helloworld case also
improves its performance to 1300MiB/s.

Because increasing the IOSIZE increases the size of the
written file proportionally (to 25GiB), this commit also
modifies the respective qsub scripts to delete the files
after executing helloworld and the warpdrive cases.  This
should minimize the risk of future ATPESC uses accidentally
exhausing the collective project quota with scratch files.